LayoutTests/discovery/get-network-services-types-expected.txt

Tests navigator.getNetworkServices() works in happy case as expected.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS navigator.getNetworkServices() first call to ok callback.
PASS srvsCDS.servicesAvailable==0 is true
PASS navigator.getNetworkServices() first call to ok callback.
PASS srvsCnn.servicesAvailable==0 is true
PASS successfullyParsed is true

TEST COMPLETE
